Refactor: PaginationButtons
